# Conflicts:
#	com.jebao.root/1.1-com.jebao.p2p.web/src/main/resources/templates/common/footer.html

+++ com.jebao.root/3.1-com.jebao.jebaodb.dao/src/main/java/com/jebao/jebaodb/dao/dao/user/TbUserDetailsDao.java
@@ -18,7 +18,7 @@ public class TbUserDetailsDao {
     @Autowired
     private TbUserDetailsMapper tbUserDetailsMapper;
 
-    public long insert(TbUserDetails record) {
+    public int insert(TbUserDetails record) {
         return tbUserDetailsMapper.insert(record);
     }
